refactor(): minor updates for next stencil
