+++ core/src/components/action-sheet/action-sheet.tsx
@@ -25,7 +25,6 @@ export class ActionSheet implements ComponentInterface, OverlayInterface {
 
   presented = false;
   animation?: any;
-  mode = getIonMode(this);
 
   @Element() el!: HTMLIonActionSheetElement;
