+++ core/src/components/loading/loading.tsx
@@ -27,7 +27,6 @@ export class Loading implements ComponentInterface, OverlayInterface {
   private durationTimeout: any;
 
   presented = false;
-  mode = getIonMode(this);
 
   @Element() el!: HTMLIonLoadingElement;
