+++ core/src/components/menu/menu-interface.ts
@@ -2,7 +2,6 @@ import { Animation, Side } from '../../interface';
 
 export interface MenuI {
   el: HTMLIonMenuElement;
-  mode: string;
   side: Side;
   menuId?: string;
   disabled: boolean;
