+++ core/src/components/router/utils/interface.ts
@@ -2,7 +2,6 @@ import { ComponentProps } from '../../../interface';
 
 export interface HTMLStencilElement extends HTMLElement {
   componentOnReady(): Promise<this>;
-  forceUpdate(): void;
 }
 
 export interface NavOutlet {
