+++ core/src/components/toast/toast.tsx
@@ -27,7 +27,6 @@ export class Toast implements ComponentInterface, OverlayInterface {
   private durationTimeout: any;
 
   presented = false;
-  mode = getIonMode(this);
 
   @Element() el!: HTMLIonToastElement;
