+++ core/src/components/toolbar/toolbar.tsx
@@ -1,4 +1,4 @@
-import { Component, ComponentInterface, Element, Host, Listen, Prop, h } from '@stencil/core';
+import { Component, ComponentInterface, Element, Host, Listen, Prop, h, forceUpdate } from '@stencil/core';
 
 import { getIonMode } from '../../global/ionic-global';
 import { Color, CssClassMap, StyleEventDetail } from '../../interface';
@@ -76,7 +76,7 @@ export class Toolbar implements ComponentInterface {
 
     if (hasStyleChange) {
       this.childrenStyles.set(tagName, newStyles);
-      this.el.forceUpdate();
+      forceUpdate(this);
     }
   }
