+++ core/src/utils/menu-controller/animations/reveal.ts
@@ -1,3 +1,4 @@
+import { getIonMode } from '../../../global/ionic-global';
 import { Animation, MenuI } from '../../../interface';
 import { createAnimation } from '../../animation/animation';
 
@@ -9,11 +10,11 @@ import { baseAnimation } from './base';
  * The menu itself, which is under the content, does not move.
  */
 export const menuRevealAnimation = (menu: MenuI): Animation => {
+  const mode = getIonMode(menu);
   const openedX = (menu.width * (menu.isEndSide ? -1 : 1)) + 'px';
-
   const contentOpen = createAnimation()
     .addElement(menu.contentEl!) // REVIEW
     .fromTo('transform', 'translateX(0px)', `translateX(${openedX})`);
 
-  return baseAnimation(menu.mode === 'ios').addAnimation(contentOpen);
+  return baseAnimation(mode === 'ios').addAnimation(contentOpen);
 };
