+++ core/src/utils/overlays-interface.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from '@stencil/core';
 import { HTMLStencilElement } from '@stencil/core/internal';
 
-import { AnimationBuilder, Mode } from '../interface';
+import { AnimationBuilder } from '../interface';
 
 export interface OverlayEventDetail<T = any> {
   data?: T;
@@ -9,7 +9,6 @@ export interface OverlayEventDetail<T = any> {
 }
 
 export interface OverlayInterface {
-  mode: Mode;
   el: HTMLElement;
   animated: boolean;
   keyboardClose: boolean;
